Ported comments for C89 and starting to adapt iterators for C89 style

+++ src/commons/linked_list.c
@@ -13,15 +13,15 @@
 #include <stddef.h>
 #include "linked_list.h"
 
-//Simple linked list implementation
+/* Simple linked list implementation */
 
 /*
 ** Create a linked list
 */
 
 l_list *list_create() {
-    //Create a empty list, this will be overitten when adding a
-    //Element
+    /* Create a empty list, this will be overitten when adding a */
+    /* Element */
     return NULL;
 }
 
@@ -33,7 +33,7 @@ void list_add(l_list **input_list, int element) {
 
     l_list  *tmp = malloc(sizeof (l_list));
 
-    //Check if the memory was allocated
+    /* Check if the memory was allocated */
     if (tmp == NULL) {
         printf("[Linked List] Can't allocate element");
         return;
@@ -48,11 +48,11 @@ void list_add(l_list **input_list, int element) {
 ** Length of the list
 */
 
-size_t list_length(l_list *list){
+unsigned long list_length(l_list *list){
 
-    size_t length = 0;
+    unsigned long length = 0;
     
-    //Iterate when there is a pointer 
+    /* Iterate when there is a pointer */
     while (list) {
         list = list->next_element;
         length++;
@@ -67,26 +67,26 @@ size_t list_length(l_list *list){
 
 int list_get_value(l_list *list, int pos) {
 
-    //Size of the list
+    /* Size of the list */
     int length = (int) list_length(list);
 
-    //Check if the list really exists
+    /* Check if the list really exists */
     if (list == NULL) {
         return -1;
     }
 
-    //Current position
-    size_t current_pos = 0;
+    /* Current position */
+    unsigned long current_pos = 0;
 
-    //Calculate real position - hotfix
-    size_t realpos = length - pos - 1;
+    /* Calculate real position - hotfix */
+    unsigned long realpos = length - pos - 1;
 
-    //Check OoB
+    /* Check OoB */
     if (pos > length) {
         return -1;
     }
 
-    //Iterate trought the list
+    /* Iterate trought the list */
     while (current_pos < realpos && list) {
         list = list->next_element;
         current_pos++;
@@ -101,26 +101,26 @@ int list_get_value(l_list *list, int pos) {
 
 void list_set_value(l_list *list, int pos, int value) {
 
-    //Size of the list
+    /* Size of the list */
     int length = (int) list_length(list);
 
-    //Check if the list really exists
+    /* Check if the list really exists */
     if (list == NULL) {
         return;
     }
 
-    //Current position
-    size_t current_pos = 0;
+    /* Current position */
+    unsigned long current_pos = 0;
 
-    //Calculate real position - hotfix
-    size_t realpos = length - pos - 1;
+    /* Calculate real position - hotfix */
+    unsigned long realpos = length - pos - 1;
 
-    //Check OoB
+    /* Check OoB */
     if (pos > length) {
         return;
     }
 
-    //Iterate trought the list
+    /* Iterate trought the list */
     while (current_pos < realpos && list) {
         list = list->next_element;
         current_pos++;
@@ -134,13 +134,13 @@ void list_set_value(l_list *list, int pos, int value) {
 */
 
 void list_free(l_list *list) {
-    //Head of the list
+    /* Head of the list */
     l_list *head = list;
 
-    //Selected node - start from the head
+    /* Selected node - start from the head */
     l_list *current_node = list;
 
-    //Free everything
+    /* Free everything */
     while ((current_node = head) != NULL) {
 
         head = head->next_element;
@@ -154,15 +154,15 @@ void list_free(l_list *list) {
 
 void list_print(l_list *list) {
     
-    //Get length
-    size_t length = list_length(list);
+    /* Get length */
+    unsigned long length = list_length(list);
 
-    //Check list size
+    /* Check list size */
     if (length == 0)
         return;
 
-    //Current position
-    size_t current = 0;
+    /* Current position */
+    unsigned long current = 0;
 
     while (current < length) {
         printf("%i ", list_get_value(list, current));
