+++ src/commons/linked_list.h
@@ -7,18 +7,18 @@
 ** Started on  2019 SheatNoisette
 */
 
-//Simple linked list implementation
+/* Simple linked list implementation */
 
 #ifndef LINKED_LIST_H
 #define LINKED_LIST_H
 
-//Imports
+/* Imports */
 #include <stddef.h>
 
-//----------------------------------
-// Structures
+/* ---------------------------------- */
+/*  Structures */
 
-//l_list is an alias of list_node
+/* l_list is an alias of list_node */
 typedef struct list_node l_list;
 
 struct list_node {
@@ -26,12 +26,13 @@ struct list_node {
     l_list *next_element;
 };
 
-//----------------------------------
-// Prototypes
+
+/* ---------------------------------- */
+/*  Prototypes */
 
 l_list *list_create();
 void list_add(l_list **input_list, int element);
-size_t list_length(l_list *list);
+unsigned long list_length(l_list *list);
 int list_get_value(l_list *list, int pos);
 void list_set_value(l_list *list, int pos, int value);
 void list_free(l_list *list);
