+++ src/recognition/text_reconstruction.h
@@ -12,11 +12,11 @@
 
 #include "../commons/linked_list.h"
 
-//---------------------------------------
-// Functions
+/* ---------------------------------------
+/*  Functions */
 
-size_t dt_get_average_char_size(l_list *input_pos, size_t number_chars);
-size_t dt_get_average_char_spacing(l_list *input_pos, size_t number_chars);
+unsigned long dt_get_average_char_size(l_list *input_pos, unsigned long number_chars);
+unsigned long dt_get_average_char_spacing(l_list *input_pos, unsigned long number_chars);
 l_list *deduct_text_structure (l_list *input_pos);
